Exercise 4.23: bloglist expansion, step11 Complete

+++ blog/tests/blog_api.test.js
@@ -18,6 +18,7 @@ describe( 'basicDBTest', () =>
     {
         let res = await api
             .get( '/api/blogs' )
+            .set( 'Authorization', helper.validToken )
             .expect( 200 )
             .expect( 'Content-Type', /application\/json/ )
         expect( res.body ).toHaveLength( helper.initialBlog.length )
@@ -25,9 +26,16 @@ describe( 'basicDBTest', () =>
 
     test( 'check id', async () =>
     {
-        let res = await api.get( '/api/blogs' )
-        // console.log(res.body)
+        let res = await api.get( '/api/blogs' ).set( 'Authorization', helper.validToken )
         expect( res.body[ 0 ].id ).toBeDefined()
+    } )
+    test( 'get blog without auth token', async () =>
+    {
+        await api.get( '/api/blogs' ).expect( 401 )
+    } )
+    test( 'get blog with invalid auth token', async () =>
+    {
+        await api.get( '/api/blogs' ).set( 'Authorization', helper.invalidToken ).expect( 401 )
     } )
 
 } );
@@ -43,7 +51,7 @@ describe( 'testing POST req', () =>
             likes: 2,
         }
 
-        await api.post( "/api/blogs" ).send( payload ).expect( 201 )
+        await api.post( "/api/blogs" ).send( payload ).set( 'Authorization', helper.validToken ).expect( 201 )
 
         let allBlog = await helper.blogInDB()
         expect( allBlog.length === helper.initialBlog.length + 1 )
@@ -57,7 +65,7 @@ describe( 'testing POST req', () =>
             url: "http://blog.cleancoder.com/uncle-bob/2016/05/01/TypeWars.html",
         }
 
-        let res = await api.post( '/api/blogs' ).send( payload )
+        let res = await api.post( '/api/blogs' ).send( payload ).set( 'Authorization', helper.validToken )
         expect( res.body.likes ).toBe( 0 )
     } )
 
@@ -69,7 +77,7 @@ describe( 'testing POST req', () =>
             likes: 2,
         }
 
-        await api.post( "/api/blogs" ).send( payload ).expect( 400 )
+        await api.post( "/api/blogs" ).send( payload ).set( 'Authorization', helper.validToken ).expect( 400 )
 
         payload = {
             title: "Type wars",
@@ -77,9 +85,32 @@ describe( 'testing POST req', () =>
             likes: 2,
         }
 
-        await api.post( "/api/blogs" ).send( payload ).expect( 400 )
+        await api.post( "/api/blogs" ).send( payload ).set( 'Authorization', helper.validToken ).expect( 400 )
 
     } )
+
+    test( 'creating new blog without auth token', async () =>
+    {
+        let payload = {
+            title: "Type wars",
+            author: "Robert C. Martin",
+            url: "http://blog.cleancoder.com/uncle-bob/2016/05/01/TypeWars.html",
+            likes: 2,
+        }
+        await api.post( "/api/blogs" ).send( payload ).expect( 401 )
+    } )
+
+    test( 'creating new blog with invalid token', async () =>
+    {
+        let payload = {
+            title: "Type wars",
+            author: "Robert C. Martin",
+            url: "http://blog.cleancoder.com/uncle-bob/2016/05/01/TypeWars.html",
+            likes: 2,
+        }
+        await api.post( "/api/blogs" ).send( payload ).set( 'Authorization', helper.invalidToken ).expect( 401 )
+    } )
+
 } );
 
 
@@ -90,7 +121,7 @@ describe( 'DELETE req', () =>
         let allBlogsInDB = await helper.blogInDB()
         let blogToDelete = allBlogsInDB[ 0 ]
 
-        await api.delete( `/api/blogs/${ blogToDelete.id }` ).expect( 204 )
+        await api.delete( `/api/blogs/${ blogToDelete.id }` ).set( 'Authorization', helper.validToken ).expect( 204 )
 
         let afterDeletingBlogs = await helper.blogInDB()
         expect( afterDeletingBlogs.length ).toBe( helper.initialBlog.length - 1 )
@@ -102,13 +133,29 @@ describe( 'DELETE req', () =>
     test( 'deleting blogs with invalid ID', async () =>
     {
         let invalidID = "5a3d5da59070081a82a3445"
-        await api.delete( `/api/blogs/${ invalidID }` ).expect( 400 )
+        await api.delete( `/api/blogs/${ invalidID }` ).set( 'Authorization', helper.validToken ).expect( 400 )
     } )
 
     test( 'deleting blog with ID which is not in DB', async () =>
     {
         let nonExistingId = await helper.nonExistingId()
-        await api.delete( `/api/blogs/${ nonExistingId }` ).expect( 404 )
+        await api.delete( `/api/blogs/${ nonExistingId }` ).set( 'Authorization', helper.validToken ).expect( 404 )
+    } )
+
+    test( 'deleting blogs without auth token', async () =>
+    {
+        let allBlogsInDB = await helper.blogInDB()
+        let blogToDelete = allBlogsInDB[ 0 ]
+
+        await api.delete( `/api/blogs/${ blogToDelete.id }` ).expect( 401 )
+    } )
+
+    test( 'deleting blogs with invalid auth token', async () =>
+    {
+        let allBlogsInDB = await helper.blogInDB()
+        let blogToDelete = allBlogsInDB[ 0 ]
+
+        await api.delete( `/api/blogs/${ blogToDelete.id }` ).set( 'Authorization', helper.invalidToken ).expect( 401 )
     } )
 
 } );
@@ -120,9 +167,9 @@ describe( 'PUT req', () =>
         let allBlogsInDB = await helper.blogInDB()
         let blogToUpdate = allBlogsInDB[ 0 ];
         blogToUpdate.likes = 7657691;
+        blogToUpdate.user = blogToUpdate.user.toString()
 
-        let updatedBlog = await api.put( `/api/blogs/${ blogToUpdate.id }` ).send( blogToUpdate )
-
+        let updatedBlog = await api.put( `/api/blogs/${ blogToUpdate.id }` ).send( blogToUpdate ).set( 'Authorization', helper.validToken )
         expect( blogToUpdate ).toEqual( updatedBlog.body )
 
     } )
@@ -133,7 +180,7 @@ describe( 'PUT req', () =>
         let allBlogsInDB = await helper.blogInDB()
         let blogToUpdate = allBlogsInDB[ 0 ];
         blogToUpdate.likes = 76571;
-        await api.put( `/api/blogs/${ invalidID }` ).send( blogToUpdate ).expect( 400 )
+        await api.put( `/api/blogs/${ invalidID }` ).send( blogToUpdate ).set( 'Authorization', helper.validToken ).expect( 400 )
     } )
 
     test( 'updating blog with ID which is not in DB', async () =>
@@ -142,6 +189,28 @@ describe( 'PUT req', () =>
         let allBlogsInDB = await helper.blogInDB()
         let blogToUpdate = allBlogsInDB[ 0 ];
         blogToUpdate.likes = 76576921;
-        await api.put( `/api/blogs/${ nonExistingId }` ).send( blogToUpdate ).expect( 404 )
+        await api.put( `/api/blogs/${ nonExistingId }` ).send( blogToUpdate ).set( 'Authorization', helper.validToken ).expect( 404 )
+    } )
+
+    test( 'updating a blog without auth token', async () =>
+    {
+        let allBlogsInDB = await helper.blogInDB()
+        let blogToUpdate = allBlogsInDB[ 0 ];
+        blogToUpdate.likes = 7657691;
+        blogToUpdate.user = blogToUpdate.user.toString()
+
+        await api.put( `/api/blogs/${ blogToUpdate.id }` ).send( blogToUpdate ).expect( 401 )
+
+    } )
+
+    test( 'updating a blog with invalid auth token', async () =>
+    {
+        let allBlogsInDB = await helper.blogInDB()
+        let blogToUpdate = allBlogsInDB[ 0 ];
+        blogToUpdate.likes = 7657691;
+        blogToUpdate.user = blogToUpdate.user.toString()
+
+        await api.put( `/api/blogs/${ blogToUpdate.id }` ).send( blogToUpdate ).set( 'Authorization', helper.invalidToken ).expect( 401 )
+
     } )
 } );
